# Use instance=True in create_autospec