--- a/docs/release_notes/next/dev-2433-autospec-speedup
+++ b/docs/release_notes/next/dev-2433-autospec-speedup
@@ -1,0 +1,1 @@
+#2433: Speedup autospec in tests