--- a/mantidimaging/gui/windows/operations/test/presenter_test.py
+++ b/mantidimaging/gui/windows/operations/test/presenter_test.py
@@ -25,7 +25,7 @@ from mantidimaging.core.data import ImageStack
 class FiltersWindowPresenterTest(unittest.TestCase):
 
     def setUp(self) -> None:
-        self.main_window = mock.create_autospec(MainWindowView)
+        self.main_window = mock.create_autospec(MainWindowView, instance=True)
         self.view = mock.MagicMock()
         self.presenter = FiltersWindowPresenter(self.view, self.main_window)
         self.presenter.model.filter_widget_kwargs = {"roi_field": None}